Create board ui movement
This method returns a highlighted area representing the given index.

For example, if given 0 it will return a highlight on the Go tile.

It relies on provider to update the current tile being higlighted.

It also uses a switch statement, which could be replaced by a series of if statements. This would clean up the space vertically, not much difference functionality wise.