Update get board method
Returns whole board as a combination of rows and columns.

Note that the board is a stack and at the bottom lies the get board portion. It is a static representation of the board. Which is why most of the elements are not added programmatically and result in many lines of styling code.

This is not the case with the movement portion of the board which is displayed on top of the static board.