Update entire board representation
Board is divided in 3 columns, left center and right. The board is made using expanded with flex widgets. This is to ensure proportions stay constants on screen rebuild. For example, in web browser it will be able to grow and shrink to almost any desirable size.

See previous commit for a better understanding.